feat!: support local mathlingua.jar
The extension will use `bin/mathlingua.jar` in the workspace
directory if the file exists.  Otherwise, it will use the
`mathlingua.jar` bundled with the extension.

Also, the `mathlingua.duplicates` and `mathlingua.undefined`
commands are no longer supported because they are handled by
`mlg check` whenever a mathlingua file is saved.